requirements.txt
streamlit==1.15.2
plotly==5.11.0
pandas==1.5.2
numpy==1.23.5
folium==0.13.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
haversine==2.7.0
streamlit-folium==0.7.0
Pillow==9.3.0
inflection==0.5.1